refactor: folder structure inside tests folder
