+++ tests/unit-tests/fights/fight-model.test.ts
@@ -1,8 +1,8 @@
 import { expect } from 'chai';
 import mongoose from 'mongoose';
 import Sinon from 'sinon';
-import { IFight } from '../../interfaces';
-import Fight from '../../models/fight-model';
+import { IFight } from '../../../interfaces';
+import Fight from '../../../models/fight-model';
 
 describe('Tests Fight Model', () => {
   const fightMock: IFight = {
